requirements/pip.txt
astroid==1.6.6
backports.functools-lru-cache==1.6.6
colorama==0.4.6
configparser==4.0.2
Django==1.11.29
django-tastypie==0.14.7
enum34==1.1.10
futures==3.4.0
isort==4.3.21
lazy-object-proxy==1.6.0
mccabe==0.6.1
python-dateutil==2.9.0.post0
python-mimeparse==1.6.0
pytz==2024.1
singledispatch==3.7.0
six==1.16.0
wrapt==1.15.0